# docs/reference/openapi.yaml
openapi: 3.0.0
info:
  title: Allo-Scrapper REST API
  description: |
    Theater showtime aggregator API for scraping and managing theater data.
    
    **Authentication**: All protected endpoints require a JWT token in the `Authorization: Bearer <token>` header.
    
    **Rate Limiting**: API enforces rate limits per endpoint group:
    - Public endpoints: 100 requests/minute
    - Protected endpoints: 60 requests/minute
    - Scraper endpoints: 10 requests/minute
  version: 3.0.0
  contact:
    name: Allo-Scrapper Project
    url: https://github.com/phbassin/allo-scrapper
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.example.com/api
    description: Production server (replace with your domain)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: admin
        role_id:
          type: integer
          example: 1
        role_name:
          type: string
          example: admin
        is_system_role:
          type: boolean
          example: true
        permissions:
          type: array
          items:
            type: string
          example: [users:list, users:create, scraper:trigger]
...
    Theater:
      type: object
      properties:
        id:
          type: string
          example: c12345
        name:
          type: string
          example: UGC Ciné Cité Paris
        url:
          type: string
          example: https://www.allocine.fr/seance/salle_gen_csalle=C12345.html
        source:
          type: string
          enum: [allocine, ugc, pathe]
          example: allocine
        created_at:
          type: string
          format: date-time
          example: 2025-03-18T10:30:00Z
        updated_at:
          type: string
          format: date-time
          example: 2025-03-18T15:45:00Z
    CreateTheaterRequest:
      type: object
      required:
        - name
        - url
        - source
      properties:
        name:
          type: string
          example: New Theater
        url:
          type: string
          example: https://www.allocine.fr/seance/salle_gen_csalle=C99999.html
        source:
          type: string
          enum: [allocine, ugc, pathe]
          example: allocine

    # Movie
    Movie:
      type: object
      properties:
        id:
          type: integer
          example: 12345
        title:
          type: string
          example: Dune Part Two
        release_date:
          type: string
          format: date
          example: 2025-02-26
        duration_minutes:
          type: integer
          example: 166
        genres:
          type: array
          items:
            type: string
          example: [Science-Fiction, Adventure, Drama]
        synopsys:
          type: string
          example: The second part of the epic adaptation...

    # Schedule
...
    ScrapeReport:
      type: object
      properties:
        id:
          type: integer
          example: 42
        status:
          type: string
          enum: [pending, running, success, partial_success, failed]
          example: success
        triggerType:
          type: string
          enum: [manual, cron, api]
          example: manual
        created_at:
          type: string
          format: date-time
          example: 2025-03-18T10:00:00Z
        completed_at:
          type: string
          format: date-time
          example: 2025-03-18T10:15:00Z
        total_theaters:
          type: integer
          example: 50
        successful_theaters:
          type: integer
          example: 48
        failed_theaters:
          type: integer
          example: 2
        total_movies_scraped:
          type: integer
          example: 1245
        total_showtimes_scraped:
          type: integer
          example: 5432
        errors:
          type: array
          items:
            type: object
            properties:
              theater_name:
                type: string
              error:
                type: string

    # Settings
    AppSettings:
      type: object
      properties:
        id:
          type: integer
          example: 1
        app_name:
          type: string
          example: Allo-Scrapper
        app_logo_url:
          type: string
          example: data:image/png;base64,...
        primary_color:
          type: string
          example: "#0066cc"
        secondary_color:
          type: string
          example: "#00cc99"
        footer_text:
          type: string
          example: "© 2025 Allo-Scrapper"

    # Role
    Role:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: admin
        description:
          type: string
          example: Administrator with full access
        is_system:
          type: boolean
          example: true
        permissions:
          type: array
          items:
            type: string
          example: [users:list, users:create, scraper:trigger]

    # System
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Theater with ID c99999 not found
        statusCode:
          type: integer
          example: 404
  responses:
    UnauthorizedError:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ConflictError:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
  - BearerAuth: []
paths:
...
  /theaters:
    get:
      summary: List All Theaters
      tags: [Theaters]
      security: []
      description: Get list of all theaters with pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: search
          in: query
          schema:
            type: string
          description: Search theater name
      responses:
        '200':
          description: Theaters list
          content:
            application/json:
              schema:
                type: object
                properties:
                  theaters:
                    type: array
                    items:
                      $ref: '#/components/schemas/Theater'
                  total:
                    type: integer
                    example: 150
                  page:
                    type: integer
                    example: 1
    post:
      summary: Create Theater
      tags: [Theaters]
      description: Add a new theater to the system (requires `theaters:create` permission)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTheaterRequest'
      responses:
        '201':
          description: Theater created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Theater'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /theaters/{id}:
    get:
      summary: Get Theater Details
      tags: [Theaters]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: c12345
      responses:
        '200':
          description: Theater details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Theater'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      summary: Update Theater
      tags: [Theaters]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                url:
                  type: string
      responses:
        '200':
          description: Theater updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Theater'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Delete Theater
      tags: [Theaters]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Theater deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /movies:
    get:
      summary: List Movies
      tags: [Movies]
      security: []
      description: Get movies with optional filtering and pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: theater_id
          in: query
          schema:
            type: string
          description: Filter by theater ID
      responses:
        '200':
          description: Movies list
          content:
            application/json:
              schema:
                type: object
                properties:
                  movies:
                    type: array
                    items:
                      $ref: '#/components/schemas/Movie'
                  total:
                    type: integer
...
  /movies/{id}:
    get:
      summary: Get Movie Details
      tags: [Movies]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Movie details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /scraper/reports:
    get:
      summary: List Scrape Reports
      tags: [Scraper]
      description: Get list of all scraping reports (requires `reports:list` permission)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Scrape reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScrapeReport'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /settings:
    get:
      summary: Get Public Settings
      tags: [Settings]
      security: []
      description: Get public application settings (branding, colors, etc)
      responses:
        '200':
          description: Public settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSettings'
    put:
      summary: Update Settings
      tags: [Settings]
      description: Update application settings (admin only, requires `settings:update` permission)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                app_name:
                  type: string
                primary_color:
                  type: string
                  format: color
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSettings'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /roles:
    get:
      summary: List All Roles
      tags: [Roles]
      description: Get list of all roles (requires `roles:list` permission)
      responses:
        '200':
          description: Roles list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      summary: Create Role
      tags: [Roles]
      description: Create a new custom role (requires `roles:create` permission)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                permissions:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /roles/{id}:
    get:
      summary: Get Role Details
      tags: [Roles]
      description: Get details of a specific role (requires `roles:read` permission)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Role details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      summary: Update Role
      tags: [Roles]
      description: Update role and permissions (requires `roles:update` permission)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                permissions:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Delete Role
      tags: [Roles]
      description: Delete a custom role (requires `roles:delete` permission, cannot delete system roles)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Role deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users:
    get:
      summary: List Users
      tags: [Users]
      description: Get list of users with pagination (requires `users:list` permission)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/{id}:
    get:
      summary: Get User Details
      tags: [Users]
      description: Get details of a specific user (requires `users:read` permission)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      summary: Update User Role
      tags: [Users]
      description: Change user's role (requires `users:update` permission)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role_id
              properties:
                role_id:
                  type: integer
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Delete User
      tags: [Users]
      description: Delete a user account (requires `users:delete` permission, cannot delete last admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...